fix: 가격 데이터타입을 double로 변경

+++ src/main/java/com/kusitms/jipbap/food/FoodOption.java
@@ -23,8 +23,7 @@ public class FoodOption {
 
     private String name;
 
-    private Long dollarPrice;
-
-    private Long canadaPrice;
+    private Double dollarPrice;
+    private Double canadaPrice;
 
 }
